requirements.txt
pytest~=7.4.2
allure-pytest
selene==2.0.0rc4